feat/gas-refund: move files around to ease future pull
