+++ src/service/transaction-fees-indexer/index.ts
@@ -12,11 +12,11 @@ import { computeMerkleData } from './merkle-tree';
 import { fetchDailyPSPChainCurrencyRate } from './psp-chaincurrency-pricing';
 import { computeAccumulatedTxFeesByAddress } from './transaction-fees';
 import { fetchPSPStakes } from './staking';
-import { Claimable, HistoricalPrice, TxFeesByAddress, PSPStakesByAddress, UpdateCompletedEpochData, MerkleData, MerkleTreeDataByChain } from './types';
-import { EpochGasRefund } from '../../models/EpochGasRefund';
+import { Claimable, HistoricalPrice, TxFeesByAddress, MerkleTreeDataByChain } from './types';
 import Database from '../../database';
 
 import BigNumber from 'bignumber.js';
+import { writeCompletedEpochData } from './db-persistance';
 
 const logger = global.LOGGER('GRP');
 
@@ -105,67 +105,7 @@ async function reduceGasRefundByAddressAllChains(
   return gasRefundByAddressByChain;
 }
 
-const writeCompletedEpochData = async (merkleTreeDataByChain: MerkleTreeDataByChain, pspStakesByAddress: PSPStakesByAddress) => {
 
-  /*
-  epoch: number                   merkleTreeDataByChain.[chainId].root.epoch
-  address: string                 merkleTreeDataByChain.[chainId].leaves[].address
-  chainId: string                 merkleTreeDataByChain.[chainId].leaves[].amount
-
-  totalStakeAmountPSP: string     pspStakesByAddress[address]
-  refundedAmountPSP: string       merkleTreeDataByChain.[chainId].root.totalAmount
-  merkleProofs: string[]          merkleTreeDataByChain.[chainId].leaves[].merkleProofs
-  merkleRoot: string              merkleTreeDataByChain.[chainId].root.merkleRoot
-  */
-
-  const epochDataToUpdate: UpdateCompletedEpochData[] = Object
-  .keys(merkleTreeDataByChain)
-  .map((chainId) => {
-    const merkleTreeDataForChain = merkleTreeDataByChain[+chainId]
-    // because `computeMerkleData` can return null
-    if (!merkleTreeDataForChain) {
-      return []
-    }
-    const { root: { epoch, totalAmount, merkleRoot }, leaves } = merkleTreeDataForChain
-
-    const addresses = leaves.map((leaf: MerkleData) => ({
-      epoch,
-      address: leaf.address,
-      chainId,
-
-      totalStakeAmountPSP: pspStakesByAddress[leaf.address].toString(), // todo: make safe
-      refundedAmountPSP: totalAmount,
-      merkleProofs: leaf.merkleProofs,
-      merkleRoot,
-    }))
-    return addresses
-  })
-  // lastly flatten the array (of chain specific arrays)
-  .reduce((buildingArray, array) => buildingArray.concat(array), [])
-
-
-  // todo: bulk upsert epoch data once models are defined
-  for (let i = 0; i < epochDataToUpdate.length; i++) {
-    const endEpochData = epochDataToUpdate[i];
-
-    // key
-    const { epoch, address, chainId } = endEpochData
-    // update
-    const { totalStakeAmountPSP, refundedAmountPSP, merkleProofs, merkleRoot } = endEpochData
-
-    const row = await EpochGasRefund.findOne({ where: { epoch, address, chainId }})
-
-    await EpochGasRefund.update(
-      {
-        totalStakeAmountPSP, refundedAmountPSP, merkleProofs, merkleRoot
-      },
-      {
-        where: { epoch, address, chainId}
-      }
-    )
-  }
-
-}
 
 async function computeMerkleTreeDataAllChains(
   claimableAmountsByChain: {
@@ -237,10 +177,4 @@ export async function start(epochNum: number) {
 
 }
 
-// todo: delete later - just created while developing
-const seedDB = async () => {
-  await Database.connectAndSync()
-}
-
 start(epochNum);
-// seedDB();
